leadgen/leadgen/backend_fastapi/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
alembic==1.13.1
asyncpg==0.29.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[argon2]
argon2-cffi
python-multipart==0.0.6
requests==2.31.0
groq>=0.9.0
email-validator
pandas==2.2.0
greenlet
openpyxl
python-socketio[asgi]